Working program logic
Known bug: While being invited and inviting the opponent and the same time, two different responses will lead the program to ambiguous state. For example when player 1 and player 2 invited each other at the same time, and one of them chose to play the game, and the other player chose not to.